planet rotation completed, window wedges are drawn
git-svn-id: http://phenomena-client.googlecode.com/svn/trunk/HelioroomTeacherDashboard@45 b2d1b222-19ed-11df-8faf-b330578705e8

+++ src/ltg/phenomena/helioroom/Degree.java
@@ -11,15 +11,15 @@ package ltg.phenomena.helioroom;
 public class Degree {
 	
 	// Always between 0 and 359
-	private float value = -1;
+	private double value = -1;
 	
 	
-	public Degree(float value) {
+	public Degree(double value) {
 		this.value = normalize(value);
 	}
 	
 	
-	private float normalize(float v) {
+	private double normalize(double v) {
 		// Normalize it
 		v = v % 360;
 		// Change sign if necessary
@@ -62,7 +62,7 @@ public class Degree {
 	}
 	
 	
-	public float getValue() {
+	public double getValue() {
 		return value;
 	}
